修改Java,postman请求成功

+++ src/main/java/com/example/uninews_admin/model/Article.java
@@ -7,13 +7,12 @@ import java.util.Date;
 @Entity
 @Table(name = "article")
 public class Article {
-
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private int id;
 
     @Column(name = "university_id")
-    private Long universityId;
+    private int universityId;
 
     private String title;
